emo_web/requirements.txt
Django>=5.1.7
djangorestframework
python-dotenv
chardet
requests
openai
